+++ src/pages/Team/styles.js
@@ -1,4 +1,4 @@
-import { makeStyles } from "@material-ui/styles";
+import { makeStyles } from "@mui/styles";
 
 const useStyles = makeStyles((theme) => ({
   team_fontNunito: {
@@ -12,7 +12,7 @@ const useStyles = makeStyles((theme) => ({
   },
   team_hrline: {
     width: "160px",
-    height: "8px",
+    height: "5px",
     margin: "0",
   },
   team_conainerLeft: {
@@ -57,6 +57,7 @@ const useStyles = makeStyles((theme) => ({
     pointerEvents: "none",
   },
   team_descriptionPopup: {
+    paddingTop: "16px",
     pointerEvents: "all",
     boxShadow:
       "0px 4px 5px rgba(46, 46, 46, 0.06),0px 1px 10px rgba(46, 46, 46, 0.04), 0px 2px 4px rgba(46, 46, 46, 0.08)",
@@ -112,7 +113,6 @@ const useStyles = makeStyles((theme) => ({
     fontWeight: "bold",
     marginTop: "12px",
     marginBottom: "10px",
-    // marginRight: "1rem",
   },
   team_cardDescription: {
     color: "#6D6D6D",
