--- src/pages/TeamPage/style.css
@@ -1,18 +1,0 @@
-.team-info-cards-container:hover .card-details {
-  opacity: 0.5;
-}
-.team-info-cards-container .card-details:hover {
-  transform: scale(1.1);
-  opacity: 1;
-  z-index: 2;
-  box-shadow: none;
-}
-.card-details:hover .img-hover {
-  border: 3px solid #48a145;
-}
-.svg-background:hover .socialmedia-icon {
-  fill: white;
-}
-.svg-background:hover rect {
-  fill: #48a145;
-}
