+++ src/routing/index.js
@@ -24,7 +24,7 @@ import VolunteerDashboard from "../components/VolunteerDashbord/VolunteerDetails
 import VolunteerOverview from "../components/VolunteerDashbord/IndividualVolunteers";
 import StatePartnerDashboard from "../pages/StatePartner Dashboard";
 import OurStory from "../pages/OurStory";
-import TeamPage from "../pages/TeamPage";
+import Team from "../pages/Team";
 import Partner from "../pages/OurPartner";
 
 // import { ThemeProvider } from "@mui/material/styles";
@@ -62,7 +62,7 @@ const Routing = () => {
       <Route exact path={PATHS.PRIVACY_POLICY} component={PrivacyPolicy} />
       <Route exact path={PATHS.AFE} component={AFEpage} />
       <Route exact path={PATHS.OPPORTUNITIES} component={Opportunities} />
-      <Route exact path={PATHS.MERAKI_TEAM} component={TeamPage} />
+      <Route exact path={PATHS.TEAM} component={Team} />
 
       <Route exact path={PATHS.OUR_PARTNER} component={Partner} />
       <Route exact path={PATHS.OUR_STORY} component={OurStory} />
